fix: moved the navigation to its own folder
