+++ frontend/backup/main copy.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:techtags/screens/navigation_menu.dart';
+import 'package:techtags/screens/navigation/navigation_menu.dart';
 
 void main() {
   runApp(MaterialApp(
